add analysis scripts

+++ tests/conftest.py
@@ -2,6 +2,11 @@ import pytest
 from opsml_data.helpers.defaults import params
 import os
 from pathlib import Path
+from opsml_data.helpers.utils import FindPath
+import pandas as pd
+from datetime import datetime
+
+dir_path = os.path.dirname(os.path.realpath(__file__))
 
 
 def pytest_sessionfinish(session, exitstatus):
@@ -13,7 +18,10 @@ def pytest_sessionfinish(session, exitstatus):
 
     paths = [path for path in Path(os.getcwd()).rglob("*.csv")]
     for path in paths:
-        os.remove(path)
+        if "pick_time_example.csv" in path:
+            pass
+        else:
+            os.remove(path)
 
 
 @pytest.fixture(scope="session")
@@ -44,3 +52,70 @@ def mock_response():
 @pytest.fixture(scope="session")
 def gcs_url():
     return "gs://py-opsml/data/20220927155229.csv"
+
+
+@pytest.fixture(scope="session")
+def pick_predictions():
+    csv_path = FindPath().find_filepath(
+        "pick_time_example.csv",
+        dir_path,
+    )
+    df = pd.read_csv(csv_path)
+    return df["NG_ORDER_ID"].to_list(), df["PREDICTIONS"].to_list()
+
+
+@pytest.fixture(scope="session")
+def unique_id():
+    id_ = str(datetime.now().strftime("%Y%m%d%H%M%S"))
+    return id_
+
+
+@pytest.fixture(scope="session")
+def sf_schema():
+    return "data_science"
+
+
+@pytest.fixture(scope="session")
+def df_columns():
+    return [
+        "ng_order_id",
+        "checkout_time",
+        "delivery_time",
+        "pick_time",
+        "drop_time",
+        "drive_time",
+        "wait_time",
+    ]
+
+
+@pytest.fixture(scope="session")
+def bundle_query():
+    query = """
+        SELECT
+        TIME_BUNDLE_ID,
+        DROP_OFF_TIME/60 AS ACTUAL,
+        (NBR_ADDRESSES*3.8531) -0.0415 AS DROP_TIME
+        FROM DATA_SCIENCE.OPSML_FP_BUNDLES_TIME_ACTUALS
+        WHERE DROP_OFF_EVAL_FLG = 1
+        AND DROP_OFF_EVAL_OUTLIER = 0
+        """
+    return query
+
+
+@pytest.fixture(scope="session")
+def order_query():
+    query = """
+        SELECT 
+        NG_ORDER_ID,
+        (3.8531) -0.0415 AS DROP_TIME,
+        NULL AS CHECKOUT_TIME,
+        NULL AS DELIVERY_TIME,
+        NULL AS DRIVE_TIME,
+        NULL AS WAIT_TIME,
+        NULL AS PICK_TIME
+        FROM DATA_SCIENCE.OPSML_FP_ORDERS_TIME_ACTUALS
+        WHERE BUNDLE_TYPE = 'TARP'
+        AND DROP_OFF_EVAL_FLG = 1
+        AND DROP_OFF_EVAL_OUTLIER = 0
+        """
+    return query
